rename enrichDocumentWithUnitex to enrichDocumentWithTerms
