+++ packages/react-tei/src/SidePanel/enrichmentTerm/EnrichmentTermSection.spec.tsx
@@ -4,7 +4,7 @@ import { render } from "vitest-browser-react";
 import { DocumentContextProvider } from "../../DocumentContextProvider";
 import { I18nProvider } from "../../i18n/I18nProvider";
 import { TestDocumentNavigationContextProvider } from "../../navigation/TestDocumentNavigationContextProvider";
-import type { TermStatistic } from "../../unitex/parseUnitexEnrichment";
+import type { TermStatistic } from "../../termEnrichment/parseUnitexEnrichment";
 import { EnrichmentTermSection } from "./EnrichmentTermSection";
 import type { EnrichmentTermAnnotationBlockType } from "./enrichmentTermAnnotationBlocks";
