+++ packages/react-tei/src/termEnrichment/createSegmentFromBoundary.ts
