+++ packages/react-tei/src/termEnrichment/propagateGroupsToSubTerm.spec.ts
