+++ packages/react-tei/src/termEnrichment/propagateGroupsToSubTerm.ts
