+++ packages/react-tei/src/termEnrichment/splitOverlappingTermsIntoSegments.ts
