+++ packages/react-tei/src/termEnrichment/splitTermIntoSegments.spec.ts
