+++ packages/react-tei/src/termEnrichment/splitTermIntoSegments.ts
